[FABRIC 1.17.1] Huge start script update
This will now download the Fabric loader, which will download the Minecraft server.jar
The file has the ability to download any version of the game, loader, and/or installer, making it extremely easy for me (or anyone) to update/downgrade Fabric and/or Minecraft.

This may soon lead to the ability to download ALL necessary files (config files, mods, etc...) and being able to choose which template you want to use, the game version, the loader version, and the installer version. This means that in the future, the templates could be shipped all in ONE batch file.