Replaced `PluginChecker` with `PluginDownloader`
PluginDownloader will automatically download Spigot plugins using Spiget. If the plugin isn't available on Spigot (or the download fails), it will give the user a link to manually install the plugin.
This should hopefully make it a lot easier to set-up Spigot servers (I'll be looking into similar solutions for the Fabric templates too).